fcntl(): fix get and set of file status flags
An application may change the file status flags via fcntl() after
creating a given file descriptor, so syscalls that act on an open
file descriptor need to be able to access the current value of the
status flags, which may be different from the flags specified when
the file descriptor was created. This is necessary for example when
an application switches between blocking and non-blocking
operation.

With this change, the flags member variable of fdesc is used to
store both the file descriptor flags (i.e. O_CLOEXEC) and the file
status flags. These two sets of flags are modified or returned via
the appropriate commands (*FD for file descriptor flags, *FL for
file status flags).

Syscalls that create a new file descriptor need to properly
initialize the flags.