TCP sockets: reset POLLOUT condition when send buffer is filled up
Analogously to what is done in sock_read_bh(), where the POLLIN
condition is reset when the socket runs out of incoming data,
socket_write_tcp_bh() should reset the POLLOUT confition when the
data being written fills up the TCP send buffer, otherwise an
application that polls on the socket for POLLOUT events may not be
notified when the POLLOUT condition is verified again (i.e. when
the TCP send buffer is not full anymore).