openat/mkdirat: fix handling of absolute path names
In all the *at() filesystem functions that take as first argument
a file descriptor and as second argument a path name, the file
descriptor should be ignored when the path name is absolute.
A new macro called resolve_dir() is being defined to implement the
above and other common functionalities.