Buffer: fix writing to a previously consumed buffer
When writing to a buffer, to retrieve the buffer location to be
written we cannot call buffer_ref() with the value of the "end"
member field of the buffer struct as offset argument, because this
doesn't work if the buffer has already been (at least partially)
consumed (i.e. if the "start" member field is non-zero). As offset
value, we should use the return value of buffer_length() instead.

A few unit tests are being added to catch this sort of bugs.