syscalls: implement unlink, unlinkat and rmdir
To remove an entry from the filesystem, the file name symbol
followed by a "null" tuple is inserted in the filesystem log; this
null tuple is encoded as a zero-length buffer; when decoding the
log, if such a null tuple is encountered, the corresponding entry
in the "children" tuple of the parent folder is removed.
In order to support the removal of an entry from a tuple, the
tagged allocator defined in boot/stage2.c (from whose heap tuple
entries are allocated) needs a dealloc function, which is
implemented as part of this commit.

Closes issue #383.