syscalls: implement times(2)
The Unix process struct has new member fields to keep track of
whether a process is executing in user or system context, and the
total CPU time spent in each of the two contexts. A new set of
functions has been defined to update these fields and to retrieve
them from the times() syscall implementation function.

One of the existing runtime tests has been enhanced to test the
times() syscall.

Closes issue #805.