unix/syscall.c: file_check(): avoid logging tuple contents
If file_check() is called for a file descriptor corresponfing to a
directory, its filesystem tuple contains child tuples for the . and
.. directory entries, and if print_tuple() is called on such tuple
it gets stuck in an infinite loop (it happens for example when
running InfluxDB with '-d', see issue #565).
With this commit, file_check() calls thread_log() to print the file
struct pointer (like other filesystem fdesc callbacks) instead of
the tuple contents.