epoll_ctl(): return EPERM if target fd does not support epoll.
The epoll_ctl man page says that the EPERM error should be returned
when "the target file fd does not support epoll. This error can
occur if fd refers to, for example, a regular file or a directory".

This commit partially fixes the timeout error when trying to insert
a record into InfluxDB, see issue #565.