blockq_check(): return -EAGAIN when action cannot be enqueued
In a future commit we are going to add a new feature to
blockq_check(), and the infinity return value will not mean an
error condition anymore. Since all callers of blockq_check() return
-EAGAIN when the block queue action cannot be enqueued, this error
value is now returned directly by blockq_check(), so the callers
don't need to check for infinity anymore.
A side effect of this change is that when blockq_check() returns an
error the caller cannot determine whether the block queue action
has been applied. For this reason, recvmsg() now deallocates its
temporary buffer when blockq_check() returns (and the corresponding
bottom half deallocates the buffer only if the "blocked" argument
is true).