Network sockets: use generic I/O completion in bottom halves
A new closure type called io_completion is being defined; this is
used in bottom half functions as a generic completion handler; the
convention for this handler is that it is called only in bottom
half context, i.e. only if an I/O operation caused the calling
thread to block.
A helper function called syscall_io_complete is also being defined;
this function does the typical stuff to be done when a syscall is
completed after having blocked the calling thread.

This commit is in preparation for the next commit, where I/O
completions will be passed to the read/write routines associated
to a file descriptor.