file I/O: add completion handler to read() and write() routines
This allows I/O routines to be called not only directly from a
syscall, but also from other contexts, e.g. a bottom half.
3 new parameters have been added to the I/O routines:
- t: indicates the thread to which the result of execution of the
routine should be reported (which may not coincide with the current
thread)
- bh: boolean value indicating if the routine is being called from
a bottom half context; this information is used when the I/O
operation cannot complete synchronously: if the current context is
a bottom half contex the thread t is not put to sleep (because it's
already sleeping) and the infinity value is returned to the caller,
otherwise the behavior continues to be as it was before
- completion: I/O completion handler that is called at the end of
the I/O operations, but only if these operations completed
asynchronously

In order for I/O routines using the blockq API to be able to safely
call the I/O completion handler (which may in theory trigger
another I/O operation that in turn uses the blockq API), a new
function called blockq_set_completion() has been defined: this
function allows a bottom half to schedule immediate execution of
a completion handler as soon as the bottom half action has been
removed from the block queue. When using blockq_set_completion(),
the completion handler is called without holding the blockq lock
(which currently doesn't exist, but will be required when thread
execution runs with interrupts enabled), so it can safely call I/O
routines, even those that result in a new action to be added to
the same blockq.

In order for I/O routines using the filesystem API to be able to
safely call the I/O completion handler (which may in theory
trigger another I/O operation that in turn uses the filesystem
API), log_write_completion() in tfs/tlog.c uses the
vector_detete() function to remove elements from the completions
vector, instead of zeroing out the vector after applying all the
completions (which would mess up things if any of these
completions pushed a new completion to the vector).

These new functionalities will allow syscalls such as readv and
writev to work with any file descriptor, not just regular files.