vector_get(): add check against negative offset
vector_get() had a check for offset values exceeding
(<vector-len> - 1), but was missing a check for offset values less
than 0. Note that a check like (base < v->start) cannot be used
because base is an unsigned variable.

A test case is being added to the vector unit tests to exercise
out of bounds access on a non-empty vector.