socket read: fix handling of chained pbuf structures
When a network packet is received on a socket, it may be delivered
by lwIP as a chain of struct pbuf elements, where the link from an
element to the next is given by the "next" member of struct pbuf.
The sock_read_bh() function was handling only the first pbuf
structure of a chain, thereby discarding the rest of data in
packets with more than one chain element.
This commit fixes the above issue, and in addition fixes another
issue in sock_read_bh() where packets received from a datagram
socket whose data does not fit in the reception buffer were not
deallocated after a partial read: the correct behavior for a
partially read datagram is to discard the data that could not fit
in the reception buffer.

Closes #881