TCP sockets: do not use lwIP PCB after a fatal error
When a fatal error occurs on a TCP connected socket, lwIP calls the
error callback supplied via tcp_err() and deallocates the PCB
structure. Thus, the PCB cannot be accessed anymore after the error
callback; to ensure this, in the callback we set to NULL the
info.tcp.lw member of the sock structure, and in socket_close() we
check for a NULL value.