Log error when clone() is passed a null child stack.
A clone with a null child stack is equivalent to a fork(), which we do
not want to support. clone() correctly aborted upon detecting that it
got a null stack, but didn't explicitly log any error, making this
situation pretty hard to spot when debugging.

Fixed by explicitly logging an error before returning.

Signed-off-by: Alexandru Lazar <alazar@startmail.com>