sock_read_bh(): fix behavior when input data size == buffer size
This fixes a regression introduced in commit
b900ee4f8525759aebabf93180cec600a13df1f0 (pull request #887).
When the size of available input data for a socket coincides with
the size of the reception buffer passed to the read() routine,
the packet buffer structure containing the input data must be
removed from the input queue and deallocated, since its contents
have been completely consumed; otherwise, a zero-length packet
buffer remains in the input queue after the call to the read()
routine, and this causes a subsequent call to read() to return 0.

Closes #891.