epoll_wait(): fix handling of closed file descriptors
If a file descriptor previously added to an epoll instance is
closed, the NOTIFY_EVENTS_RELEASE event is generated, which causes
epoll_wait_notify() to unregister the file descriptor from the
epoll instance. If epoll_wait() is then called afterwards, it
finds the unregistered file descriptor in the fds bitmap, thus an
assert(efd->registered) call at this point would fail.
This assert failure has been observed when running Nginx.

This commit removes the assert() call from epoll_wait(), and adds a
check for efd->registered before the call to fdesc_check().