select_notify(): fix handling of empty event mask
select_notify(), like the other event handlers in src/unix/poll.c,
may be called when there are no events to report; for example, when
select() is called, the event handler is called to check for any
active events on each file descriptor in the select() list. Thus,
select_notify() should exit gracefully when there are no active
events, instead of triggering an assertion failure.

Closes #902.