Add basic unit testing for random
Part of #184.

Add a basic set of unit tests for random(). Currently, this relies on
ent and performs only the most trivial tests.

While this is not sufficient to exhaustively test the quality of our
implementation, it should be reasonably sufficient to expose the most
trivial bugs that we are likely to encounter.

Signed-off-by: Alexandru Lazar <alazar@startmail.com>