clock_gettime(2): properly handle monotonic clock
The now() function returns the wall clock time, so it will be
subject to discontinuities if the system time is changed (when
support for changing the system time is actually implemented), and
for this reason it should not be used as time source in
clock_gettime() in case the requested clock is a monotonic clock.
This commit uses uptime() as time source in such cases.
In addition, the CLOCK_PROCESS_CPUTIME_ID clock type is now
implemented as well.

Closes #357.