sendfile(2): add support for any input file type
The current sendfile() implementation, which is limited to
supporting only regular files as input, is broken because
filesystem tuples are not in the current->p->process_root table,
so table_find(current->p->process_root, infile->n) always returns
false and as a result sendfile() always returns -ENOSYS.
Instead of fixing support for regular files as input to sendfile(),
this commit removes the limitation of the current implementation,
and adds support for reading any file type.

sendfile(2) is needed by Lwan (ticket #796).