ap initialization and per-cpu structures (#1061)
This is the union of the mpinit branch, the addition of per-CPU data structures (to which the various stacks and contexts have been migrated for the most part) and other ingredients for the ongoing SMP work.

This seemed as good a point as any for checkpointing into the mainline; SMP is disabled by default, and none of the AP initialization code is called. However, AP bringup can be tested out by adding "-smp n" to the qemu command line (where n is the total number of cpus) and defining SMP_TEST in the build.

An unrelated issue was found in the delivery of sigsegv within the unix fault handler. The fault handler was delivering the signal and then calling thead_yield - and in turn runloop - from within an interrupt context. This has been changed to allow the fault handler to exit gracefully while being assured that the signal will be delivered on return from exception handling.

Please note that an issue arose in the macos/clang build which leads to incorrect construction of IDT tables. A hack (making write_idt noinline) was added as a temporary workaround, but no cause has been isolated as of this point. See #1060 for details.

Co-authored-by: convolvatron <yuri@tenuki.org>