new lock-free queue (#1066)
This is an implementation of a lock-free, multiple producer / multiple consumer ring queue with single producer (enqueue_single) and single consumer (dequeue_single) variants. It replaces src/x86_64/queue.c. The interface is largely unchanged outside of the addition of the single variants, however a return value of INVALID_ADDRESS is now used to indicate an empty queue from a call to dequeue, dequeue_single or queue_peek, and a null value may be legitimately stored in the queue. Queues are instantiated in only one way despite use of multi or single variants; the caller must only be consistent in use of the interface. MPMC, MPSC, SPMC and SPSC combinations are all supported.

Two sets of head and tail pointers are used for both producing and consuming, and the typical ring enqueue and dequeue operations are split internally into reserve and commit phases (for the multi variants). See https://doc.dpdk.org/guides/prog_guide/ring_lib.html for a visual description of this data structure.

A pthread-based unit test is included. Each thread operates on a common queue, continuously alternating between enqueueing and dequeueing randomly-picked numbers from a global vector for a set duration (default 8 threads for 5 sec). Parity between enqueues and dequeues and conservation of the picked numbers are asserted.