TFS: simplify delete and rename functions
With this commit, file delete and rename functions do not take as
argument a full file path; instead, they take a tuple referencing
the parent of a file entry, and a reference to the file (either the
symbol for existing file entries, or the file name for file entries
to be created). This change allows us to get rid of the
fs_get_parent_child() function in tfs.c, which was duplicating code
from resolve_cstring() in filesystem.c.