File/dir creation: use resolve_cstring() for path traversal
With this commit, file and directory creation functions do not take
as argument a full file path; instead, they take a tuple
referencing the parent directory of an entry to be created, and the
name of the entry. Thus, they rely on resolve_cstring() in
filesystem.c for path traversal and related error handling, instead
of calling filesystem_mkentry() which duplicates some code from
resolve_cstring().
In addition, these functions now take as parameter an asynchronous
completion handler; this functionality is now used in the mkdir
syscall.
To cater for the needs of register_special_files(), the
functionality that previously was in filesystem_mkdir() is now in
the new function filesystem_mkdirpath().