syscalls: fstat: only cast fdesc to file for relevant fdesc types
It is not safe to cast an fdesc to a file and access file member
fields (in this case, the tuple field) unless the fdesc file
indicates that the fdesc is actually part of a file struct.

In fstat(), check the fdesc type, and set the tuple variable (which
is passed to fill_stat()) to 0 if there is no file struct (and
hence no tuple). This will allow fill_stat() to know unequivocally
whether its tuple argument is a valid tuple.