TFS: create_extent(): separate storage allocation from file update
With this change, create_extent() does not add the new extent to a
file (which is now done by the new function add_extent_to_file()),
but just allocates the extent; the new function fs_new_extent()
does everything create_extent() used to do.
This refactoring will be used in a future commit where we need to
allocate a number of extents and only add them to a file if there
is enough storage space for all of them.