TFS: add support for uninitialized extents
An uninitialized extent is one whose data has never been written to
disk; as such, when its data is being read, instead of reading from
disk the destination buffer is filled with zeros. When the data of
an uninitialized extent is being written, the extent loses its
uninitialized state, and its data is written to disk (any byte
ranges not present in the source buffer are filled with zeros).

Uninitialized extents are used to efficiently allocate disk space
for a file without having to write file data to disk.