Sockets: add generic ioctl() implementation
The new function socket_ioctl() is meant to be called from
socket-specific ioctl functions as a fallback for all ioctl
operations that either are not supported or can be handled by
generic socket code (as of now, only FIONBIO is supported).
The existing ioctl function for network sockets has been renamed
from socket_ioctl() to netsock_ioctl().

This commit includes also a fix for how to handle the FIONBIO
request: the corresponding argument must be treated as a pointer to
int instead of as an int.