Network sockets: sendmsg/recvmsg: fix handling of completions
With the current code, if a call to sendmsg() or recvmsg() returns
successfully without blocking, the completion function tries to
wake up the current thread, which is not blocked on anything; this
is being fixed by geting rid of the boolean "blocked" parameter and
using the syscall_io_complete closure instead.
In addition, with the current code, if a call to sendmsg() or
recvmsg() returns successfully without blocking,
sendmsg_complete_internal() or recvmsg_complete_internal() is
called twice; fix this by removing the calls to these functions
from sendmsg() and recvmsg(), and making sure the I/O completion is
always called from socket_write_internal(), regardless of the
return value.