Table: add table_clear() function
This function removes all the entries from a table, without
deallocating the table itself. The existing function
deallocate_table() has been refactored to call table_clear().
The table unit tests have been amended to exercise the new
function.
A clear_tuple() function has been added as well, which simply calls
table_clear().

This funcionality will be used in the next commit to clear unused
filesystem-related tuples without deallocating them, which is
needed because tuples should not be deallocated as long as they are
present in the filesystem log.