Filesystem log: fix updating of tuples
After the contents of a given tuple have been written to the TFS
log (using an "immediate" tuple encoding), if the tuple contents
are later changed the tuple should not be re-written to the log
using the immediate encoding, but a reference encoding should be
used instead. This is because when the log is read back, every time
an immediate encoding is encountered a new tuple is allocated and
inserted into the dictionary (whereas the tuple update done when
writing the log didn't create a new dictionary entry, because the
tuple was already present in the dictionary), which creates a
mismatch between the dictionary used when the log was written and
the dictionary created by the log reader. This mismatch may prevent
the log reader from referencing the correct tuples and symbols in
subsequent log entries, which breaks filesystem persistence.

To fix this issue, the below rules are now followed:
- a tuple is inserted into the TFS log using the immediate encoding
if it was not present in the dictionary, and using the reference
encoding otherwise
- a tuple present in the TFS log should never be deallocated,
because if a new tuple is allocated later on and happens to be
allocated at the same address as the previous tuple, if the new
tuple is written to the TFS log it would be written with a
reference encoding (because it is already present in the
dictionary), which is wrong because it is unrelated to the previous
tuple
- if the contents of a tuple change after it has been written to
the log, the log should be updated with a reference to the tuple
followed by the changed symbols and their values (a "null" symbol
value is used to indicate that a symbol is being removed from a
tuple)