Futex: fix wait timeout handling
If a thread invokes a futex wait operation with a timeout, and no
other threads run before the timeout expires, at timeout expiration
the current thread is still the thread that was blocked on the
futex. Thus, futex_bh() should check the blockq flags before
checking the current thread to determine whether it should block,
otherwise in the above scenario it does not wake up the waiting
thread at the timeout.