Page cache sync to storage: only write blocks with changed data
When contents of a pagecache page are modified by the application,
it's usually not necessary to sync the entire page to storage: only
the blocks with modified contents need to be synced.
This commit changes the way a page is synced to storage so that
unneeded disk writes (which make filesystem writes inefficient,
especially small writes) are avoided.