mkfs: add command line option to set minimum file size
The 's' command line option can be used to specify a file size for
the disk image being created. The size can be expressed in bytes,
KB (with k or K suffix), MB (with m or M suffix), and GB (with g or
G suffix); if not multiple of the disk block size, it is increased
to the next block size multiple. If the specified size is less than
the minimum size required to create the file system, this command
line option is ignored.
This change allows creating disk images with an arbitrarily large
free disk space.

Closes #1136.