ftruncate(2): update length member of file structure
Any change in the length of an open file done via ftruncate() must
be reflected not only in the filesystem but also in the file struct
allocated for the open file.
This commit implements the above behavior and adds a test case that
would trigger a failure if the above was not implemented.