Buffers: add buffer_set_capacity() function
This function allows to shrink or extend a buffer to a desired
capacity. If the buffer contents are larger than the desired
capacity, or the requested memory could not be allocated, the
buffer capacity is not changed.
This function returns the new capacity of the buffer.
The existing function buffer_extend() has been refactored to use
buffer_set_capacity().

A new test case has been added to the buffer unit tests to exercise
this feature.