xen network: fix handling of 16-bit IDs in TX requests
The ID assigned to a xennet_tx_buf is used, after left-shifting by
8 and ORing with a page ID, to create a Xen TX request ID. Since
Xen network TX request IDs are 16-bit values, any xennet_tx_buf ID
large than 255 overflows the 16-bit request ID, and as a result
when the corresponding response is consumed from the TX ring the
front driver associates the response to the wrong xennet_tx_buf.
This commit changes the bit shift value from 8 to 5 (since the
maximum pbuf length is 64 KB, the maximum number of pages that may
be needed for a single pbuf is 17) in order to allow up to 2048
xennet_tx_buf structures to be used, and adds checks to ensure that
the limit on the number of allocated xennet_tx_buf structures is
not exceeded.

Closes #1157.