EPOLL_CTL_ADD: fix handling of unregistered file descriptors
If a file descriptor is registered to an epoll instance, when the
file is closed the descriptor is unregistered from the epoll
instance, while the corresponding epollfd structure remains
associated to the epoll instance. If the file descriptor is later
re-used (by opening another file) and re-added to the epoll
instance, the EPOLL_CTL_ADD operation fails with -EEXIST when the
epollfd structure is found in the epoll instance, but the correct
behavior is for EPOLL_CTL_ADD not to fail. Fix this issue by
checking whether a given epollfd is registered, and only failing if
it is actually registered.
A test case that catches this issue has been added to the epoll
runtime tests.