netsock_bind(): do not set socket state to TCP_SOCK_OPEN
Setting the state to TCP_SOCK_OPEN when binding a TCP socket makes
it impossible to later connect() the socket to another address
(calling bind() and then connect() on a given socket is a perfectly
valid use case).
This commit fixes the above issue by not changing the socket state
in netsock_bind(), and instead checking the binding state via the
local_port field of the underlying lwIP structure; the same check
is also added for UDP sockets.