select_notify(): check fd sets before setting event flags
If a file descriptor on which a select() call is waiting is closed,
the EPOLLHUP event flag is set; this flag is included in both
POLLFDMASK_READ and POLLFDMASK_WRITE, thus if the select() caller
is interested in either read or write events (but not both), either
wset or rset may be NULL in the epoll_blocked kernel structure, and
as a result select_notify() could dereference a NULL pointer.
Fix the above issue by checking that a given bitmap (rset, wset or
eset) is non-NULL before accessing it.