syscalls: validate_iovec(): skip array entries with zero length
It is legal to pass to the readv and writev syscalls iovec
structures with zero length (which should be simply ignored). When
such a structure is found, the corresponding iov_base field should
not be checked for validity (it will most likely be a NULL
pointer), otherwise spurious EFAULT errors may be returned to
userspace.