Network interfaces: add loopback interface
This is needed by the Erlang runtime system, which when started in
"distributed" mode spawns the Erlang process mapper daemon and then
communicates with it via the loopback interface. This also prevents
a "Failed to find the loopback interface" NetUtil warning from
OpenJDK.
The loopback interface is assigned index number 0 in the network
interface list, thus the "real" network interface is assigned index
number 1 and then named "en1".
In order to receive packets from the loopback interface,
netif_poll() must be called by the kernel, and this is achieved via
the netsock_poll() thunk, which is put in the runqueue and then
invoked by the scheduler.