close(): add support for blocking the current thread
This commit changes the definition of close() in the fdesc struct
by adding a I/O completion, so that close() implementations are
allowed to block the current thread.
The new fdesc_get() and fdesc_put() functions are meant to be used
to increase and decrease, respectively, the reference count
associated to a given file (and to call close() when the reference
count goes to zero), in order to avoid deallocation of a given
fdesc while it's in use by another thread (example: a thread is
blocked polling a given file descriptor, and another thread closes
the file descriptor: in this case, the close() internal function
should not be called until the polling in the first thread
terminates).