blockq_check(): do not set blocked_on field when in_bh is true
If blockq_check() is being called from a "true" bottom half
context, there is no need to set the blocked_on field, because it
should have already been set in a previous call (before entering
the bottom half context).
Not setting the blocked_on field when in_bh is true allows to use
blockq functions in scenarios where the current thread does not
want to block.