Filesystem dump: add option to display filesystem as a tree
The mkfs/dump tool now accepts the '-t' command line option, which
makes the tool display the filesystem from a given image as a tree.
In the displayed tree, directory entries are blue, symbolic links
are cyan (with indication of the link target path), and regular
files are white.

The existing functionality to dump filesystem contents into a given
directory can now be invoked via the '-d <target dir>' command line
option.

Closes #564.