mkdir/mkdirat: fix directory creation when path ends with '/'
mkdir() and mkdirat() should detect if the supplied path ends with
the '/' delimiter, and strip the trailing delimiter if found. The
newly defined dirname_from_path() function implements this.

The mkdir suite of runtime tests has been modified so that it would
fail if the above was not implemented correctly.