id_heap_alloc_gte(): search for non-allocated ids starting from min
If the count argument passed to id_heap_alloc_gte() is 1, the
return value is guaranteed to be the lowest-numbered non-allocated
id starting from the min argument.
This is needed to correctly implement the semantics of dup2() and
fcntl(F_DUPFD).
Now that the dup runtime test passes, it is being added to the set
of CI-enforced tests. The fcntl runtime test has been amended with
a new test case that would have failed without this commit.

Closes #1012.