Sockets: make use of the domain field of struct sock
This field, which identifies the address family associated to a
socket, will be used in socket-related syscalls to safely
dereference structures associated to specific socket domains.
The `af` field of the netsock structure is being removed, because
it is redundant and the `domain` field of struct sock can be used
instead.