Socket syscalls: properly validate file descriptor number
Every socket-related syscall must validate the supplied file
descriptor number by not only checking that it corresponds to an
open file descriptor, but also checking the file type and, if the
file type is verified to be a socket, checking the socket domain;
otherwise, it's unsafe to dereference the netsock pointer derived
from the open file descriptor.
In addition, this allows user programs to call functions such as
getsockopt() to check whether a given file descriptor is a socket
descriptor (all socket-related syscalls must return -ENOTSOCK if
called on a non-socket file descriptor).