partition_get(): add check for MBR signature
If the MBR signature is not found, partition_get() returns NULL.
This will allow stage3 to detect whether the MBR has been loaded in
memory (the MBR may not be loaded if the boot process doesn't go
through stage1 and stage2).
The signature check can now be removed from mkfs.c and dump.c.