Stage3: add support for loading kernel symbols from ELF file
This will allow kernel symbols to be loaded even if the kernel is
booted without going through stage1 and stage2.
The kernel ELF file is read from the boot filesystem.
To optimize boot time in cases where kernel symbols are not needed,
the boot filesytem is loaded and the kernel file is read only if
the "ingest_kernel_symbols" symbol is present in the manifest file.