syscalls: ioctl: add ioctl_generic() function
This function is meant to implement all ioctl() commands that
touch fdesc structure members. It is called by the main ioctl()
function if a given fdesc does not provide a specific ioctl
implementation, and can also be called by a specific ioctl
implementation (such as the socket ioctl). This prevents code
duplication when handling generic fdesc commands in specific ioctl
implementations.