vmap_update_protections: fix flags for newly allocated ranges
This commit fixes a regression introduced in commit 5a4247ea; if an
mprotect syscall causes a new memory map range to be allocated at
the tail of an existing mapping, its protection flags should be
those of the existing mapping, rather than those being set for the
memory range passed to mprotect().
The mmap runtime test suite is also being amended with a new set of
tests that would have failed without this fix.