CircleCI: bump golang version in Docker image to Go 1.14
This is needed by the Azure SDK for Go, which requires Go 1.13 or
newer.
A go.mod file is being added to test/go/ because it is needed with
Go 1.13 or newer in order for the github.com/nanovms/ops/lepton
dependency to be resolved by "go build".
The go.sum file that is created by "go build" is being gitignored.