TFD log read: process decoded tuples recursively
There may be cases where the tuple corresponding to a new file
(i.e. a file not yet present in the TFS log) is written to the log
as part of writing or updating the contents of another tuple.
In order to correctly process the tuple of the new file when the
log is read back, log_parse_tuple() now calls a new helper
function, log_process_tuple(), which calls itself recursively if
the value found in a given tuple is another tuple.
The call to log_write() in filesystem_creat(), which had been added
to work around the above deficiency in the TFS log code, is now
superfluous and has been removed.