Special files: make /proc/self/exe persistent on storage
The previous code to create the /proc/self/exe symbolic link was
relying on the presence of the /proc/self/ directory. Unless
specified in the manifest, this directory was created as a
non-persistent filesystem entry when creating the /proc/self/maps
special file. Creating a persistent entry (in this case, a symlink)
as a child of a non-persistent directory makes the entry orphaned
when the machine is stopped, which means that at the next boot the
entry will not be linked to the filesystem; as a result, the
/proc/self/exe was being re-created at each boot, thereby wasting
storage space.
This commit creates the /proc/self/exe symling before the special
files, creating the /proc/self/ directory (as a a persistent entry)
if not found. This way, the symlink is not orphaned when the
machine is stopped, and will be found in the filesystem at the next
reboot.