null_value: avoid allocation from tuple-tagged region
"null_value" is a buffer, not a tuple, and as such it shouldn't be
allocated from the tuple-tagged memory region, otherwise if this
value is inserted into a tuple that is later destroyed,
destruct_tuple() mistakenly treats the value as a tuple.
This commit moves the allocation of null_value from init_tuples()
to init_runtime().
In addition, destruct_tuple() now checks that a buffer is different
from null_value before deallocating it (null_value is a singleton
value that can be shared by different tuples and should never be
deallocated).