create_filesystem(): do not take root tuple as function parameter
All current uses of create_filesystem() pass an empty tuple to this
function. Having the root tuple allocated internally in the TFS
code avoids the duplication and copy of tuple properties from the
internally allocated root to the caller-provided root when a TFS
log is read, and will make it easier to detect obsolete tuples when
log compaction is implemented.
For consistency, the root tuple is now deallocated in
destroy_filesystem() (by calling destruct_tuple() instead of going
manually through the filesystem entries, which leaks memory if
there are folders) when a filesystem is destroyed.