TFS log: calculate statistics on total entries and obsolete entries
The encode_tuple(), encode_eav() and decode_value() functions have
been enhanced to count the number of total entries and obsolete
entries found in a buffer (the encoding functions rely on being
called before the value associated to a given symbol is updated in
a tuple, so that the current value can be compared with the value
being encoded).
This feature is used in the TFS log code to keep stats that will be
used in a future commit to determine when the log should be
compacted.