filesystem_creat(): do not set filelength value in file tuple
The length of a file is by default set to 0 in allocate_fsfile(),
so there is no need to encode a 0-valued file length in the TFS
log. In addition, having a buffer shared between different tuples
would screw things up when files are deleted and their tuples have
to be destroyed.